Use components
