+++ lib/index.js
@@ -1,4 +1,4 @@
 import { string, element } from 'deku';
-import renderArticle from './render-article';
+import Article from './components/article';
 
-module.exports = tree => string.render(<article>{renderArticle(tree)}</article>);
+module.exports = items => string.render(<Article items={items || []} />);
